requirements.txt
tqdm==4.62.1
numpy==1.21.2
scikit-learn==0.24.2
kneed==0.7.0
sat-searh==0.3.0
geopandas==0.9.0
intake-stac==0.3.0
rioxarray==0.7.0
rasterio==1.2.6 
folium==0.12.1
matplotlib==3.4.3
earthpy==0.9.2
fpdf==1.7.2